requirements.txt
Brotli==1.0.9
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
dash==1.21.0
dash-bootstrap-components==0.13.1
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-renderer==1.9.1
dash-table==4.12.0
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.28.3
future==0.18.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
patsy==0.5.2
Pillow==8.4.0
plotly==5.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
retrying==1.3.3
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
statsmodels==0.13.1
tenacity==8.0.1
tomli==1.2.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
